test: fold engine-provisioning tests into the package engine_test.go
Merge the changelog-named *_fixes_test.go files into their package's
subject-named engine_test.go (ecs review_fixes into ecs_test.go), reusing the
existing recordingEngine helper. No behavior change; test naming now follows
the by-subject convention.
